Add files via upload
In the previous version of this MATLAB code, we utilized tables to store and manipulate the data. However, tables are not the most efficient data structure for MATLAB's vectorized operations. In this new version, we have refactored the code to use matrices, which allow better performance through MATLAB's optimized matrix math functions.

Overall, by migrating the data structure and corresponding operations to use matrices rather than tables, we have been able to boost the performance and take better advantage of MATLAB's matrix-based architecture. The code is now faster, more concise, and more idiomatic MATLAB.